requirements.txt
python-multipart>=0.0.6
fastapi>=0.109.0
uvicorn>=0.27.0.post1
python-dotenv>=1.0.1
langchain>=0.1.4
langchain-openai>=0.0.5
tiktoken>=0.5.2
pytest>=8.0.0